Renamed and moved around files and variables.
The `makefile` should now be clearer about what it expects, and where.

The SVG logo is now expected either in the `$(THEME)` directory or in the
`$(SOURCE)` directory. In the future, I'll probably re-integrate the
`slakkenhuis` style as the default style; this is in preparation for
that.

Also fixed a bug in the `index/view.js` (previously `sitemap/index.js`)
file, in which the relative linking meant that it would only have worked
in the build directory's root.