Introduce stylesheet in this repository.
I had decided to seperate the logic from the styling code that I wrote,
but on second thought, I think it makes sense to have the style present
as a default in a subdirectory of this repository, since the style and
generated code are somewhat intricately linked.

Unfortunately, I lose a lot of history by re-introducing the code in
this way (messy as the history may have been). I might decide to graft
the two repositories later.