Change index generation method.
Before, the index was generated with BeautifulSoup, completely in
Python. I'm now reverting back to generating an extensive sitemap in
JSON. This will be manipulated by `jq` into a HTML index.

This seems more transparent. I would have liked to use a proper
templating engine, but most of them are JavaScript-based or tied to a
programming language, and those that aren't (`mustache`) are a bit too
logic-less - I don't like how recursion is handled there.

This commit is still rough around the edges: files to ignore are not
properly ignored, and I'm sure there's other things I didn't think of
yet. Just want to commit with a proper message so that I don't forget
what I was doing when I return to it - which will inevitably be months
later.