Put link "back to index" in a more sensible place.
It was on the page-wide table of contents before, which doesn't make
sense.

We are using the "identical to" Unicode point &#8801;, not the "trigram
for heaven" point &#9776;. Either would result in three lines, which I
think is quite universally recognized as a "menu" button.

In general, I'm quite critical of the current trend of using only icons
rather than text or a combination - their significance is often not
obvious, and it's not as easy to remember which icon is supposed to do
what. However, since this will be the *only* icon, and only placed for
convenience to get back to the index page, I think an icon is the way to
go.