Add title-prefix styling.
A prefix can now be added to a page, so that the "collection title" of
the page can be shown without writing it in the actual title of the
page. It will be shown like so:

      prefix
      ------
    MAIN TITLE

Since the line height of the page header is now no longer known, I'm not
entirely happy with the way the index link doesn't perfectly line up
with the title anymore. Of course, this only affects the page if there
*is* a title prefix.

However, the flexbox approach does have some downsides: it only works on
modern browsers (not a big deal) and it will now no longer wrap properly
when the page gets too small (big deal). I might reconsider.

The table of contents has now also slightly changed positions since that
makes more sense: it now lines up with the first line of text, rather
than with the title.