Change title prefix styling approach.
Due to the problems mentioned with the flexbox approach in the previous
commit (7425be2670ddbfef24cf1a47e72c7c3e770cf6d9), I have changed it in
favour of an approach with an invisible placeholder to keep the
centering correct. It feels a bit hacky, so a better solution is
welcome.