Softcode links into the template.
The links to the favicon, stylesheet and index link used to be
hard-coded. Now they can be inserted or removed with the metadata. This
is especially useful for the index link, as one might not always want it
in there.

Using the --css argument for the stylesheet was challenging due to the
way pandoc canonicalizes the paths.