Only generate files reachable from the ToC.
Before, we would simply generate all HTML files for which there existed
a Markdown source. Now, we first build the table of contents, consider
which direct and indirect targets are needed, and on second run of make,
make those targets.

I'm unhappy with the fact that this requires two runs of make, but I
guess that means that it already did so before (for the indirect
targets).