Add headers in the style of my CV.
I think I prefer this to the single- and double lines I previously used.
I liked the lines because they are close to how they would be
represented in the Markdown sources, and they are 'simpler', in a way;
more monochrome; but I think this is actually easier to read.