Generate actual Makefile for indirect targets.
Before, we would generate documents and their external targets by
looking for the documents, examining their metadata and links, and doing
a second run of `make` based on a `.txt` file containing those targets
collected in the first run of `make`.

Now, we just include a `Makefile` that `make` understands it must
generate on its own.

This is a big improvement: we no longer have to call `make` twice. We
also no longer do double work when calling with `make -B`.