Simplify drafts: mark for removal, then remove.
This is a simplification of the way draft files are identified and
removed from the index.

I had been struggling with this for a while: everything I tried led to
some error or other. I then figured it's an actual bug in Debian
stable's version of `jq` (1.5 at the time of writing). Updating to 1.6
fixed most of the inexplicable errors (that I would like to document
here but have been lost in a nauseating mess of trial and error).

Unfortunately, it still leaves some undesirable behaviour. Notably, I
wish I could do something like this to remove elements:

    ( all_children | select(some_condition)) |= empty

This *sometimes* works, but not always; sometimes it leaves elements
that should be removed alone. So, instead, I am doing this:

    (( all_children | select(some_condition)) |= mark) | remove_marked

I suppose the reason that the first does not work predictably is that
we'd be generating elements --- children and siblings alike --- and
removing them as we're going along. There are no error messages, though,
so I'm not sure. I am not familiar with the internals of `jq`. Perhaps
it is a good idea to gain some insight to make this bulletproof, or
improve upstream `jq`'s error reporting.

For now, this is already far better than what I did before.

Again, the approach has not been applied across the board yet; this is
more of a proof of concept commit.