Simplify other jq operations to new way.
This simplifies a lot of operations. The simplifications are not just
cosmetic or just for code clarity: the behaviour is slightly different
too.

For example, frontmatter must now be explicitly marked; having an
`index.md` file does not cut it anymore. Moreover, a document will only
be taken into account when it is explicitly marked as publish; excluding
a directory by marking it as draft is no longer possible.

I think these simplifications are just as helpful as the obvious
increase in code quality: they force a single, general way of doing
things.