Generate Makefile entry for each individual file.
We used to just have a phony target with, as dependency, a list of all
the documents we're making, and a seperate phony target for all the
external targets referenced in those documents.

Now, *each* file has its own (possibly recipe-less) entry in the
makefile. The dependency list may be empty.

This has several advantages:

-   Tab completion on `make ...` works for individual documents now,
too.

-   Since we can differentiate between dependencies of PDFs and
dependencies of HTML files, the dependencies are now resolved
faithfully: when an external resource for a PDF file is changed, the PDF
itself is also rebuilt, but when an external resource for a HTML file is
changed, only the resource itself is rebuilt, since the HTML only links
to the resource.

-   We only need to make the CSS files that are actually needed for the
specific documents we are building.