Distinguish between target formats in metadata.
Instead of requiring a boolean `publish` entry in the metadata for
`snel` to decide whether or not to use the file as a source, we now
require a `make` entry in the metadata that contains a target extension
(or multiple target extensions). This way, we don't need to specify
`make html` or `make pdf`: the files already know. It also makes the
`jq` code simpler and more abstract, and better communicates what format
the style of a document might be intended for.

To simplify things, we also use a single default style.