Links now determine presence in index.
Before, we would need to explicitly determine whether an entry shows up
in the index. We now use the `link` instead, which is only present for
entries with a target format, so the effect is the same.

It also removes the necessity of having an `external` marker for
external links in the index.