Remove reference replacing.
I tried to rewrite the Python filter that temporarily but the references
in the metadata. However, it doesn't work, since Lua filters require us
to be much stricter about what to put in the metadata (MetaBlocks) and
what to put in the content (normal Blocks). So this doesn't work:

    local refs = pandoc.List({})

    return {
        {
            Div = function (el)
                if el.identifier == "refs" then
                    refs:extend(el.content)
                end
                return {}
            end,
            Meta = function (meta)
                meta.refs = pandoc.MetaBlocks(refs)
                return meta
            end,
        }
    }

I decided for the time being to just remove that approach completely and
follow Pandoc's standard behaviour. That's probably best in general: it
causes the least confusion to anyone who is already familiar with
Pandoc.