Rebuild next and prev pages upon changing a page.
In commit 2692b98bece152799fd7c92268a4d69b63b28ba5, I complained that
when a page gets added or changed, its next and previous pages are also
affected since they contain links to that page. This has been taken into
account: the source of those pages is a dependency of the current page.

Removing a page will probably still bring difficulties, but that is
harder to fix --- in that case, you'd just have to `touch` the next and
previous pages to signify that they should be rebuilt.

I also forgot to sort before adding the navigation links, which has been
rectified now.