Pass metavariables from index.json to recipes.
In commit 2692b98bece152799fd7c92268a4d69b63b28ba5, I mentioned that I
could perhaps eliminate the repetition for variables from the
`index.json` that I'd want to pass back to recipes. I have done that
now: we just set a per-recipe `PANDOC_ARGS` variable in the dynamic
Makefile that literally contains the `--metadata key=value` arguments.

This shorter and clearer in intent. Admittedly, it does also feel a bit
contrived in the way that we just convert objects into shell arguments,
but that was already the case --- just manually.