Add note about automatic links to dry-links.lua.
I didn't know that automatic links were a thing, but presumably they are
a better solution for most use cases for which I intended this filter.
I'm still keeping this filter because it does have some additional uses:
for example, it can put the domain automatically.

Also move `links.lua` to `dry-links.lua`, because it is a more
descriptive name.